Moved files out of subpackages.
Sorry for this mess. But I thought really hard about it and decided to
recreate the original package structure. The problem is, that with all
the new packages, a huge API break will occur. Not all of the classes
will create problems, but things like the IOHelper stuff is used in
workflows. The structure proposed by Florian was good and logical and I
think we can do this with the next Roddy release. I don't think, we will
have too many customers for the library in the near future. In regards
to the cluster lib, I think it is safe for this. The cluster lib has
some dependencies to the Tool Lib but that will not create problems.